src/requirements.txt
requests
python-dotenv
psycopg2-binary>=2.9
pytest
